pyproject.toml
[tool.poetry]
name = "satni_backend"
version = "1.2.5"
description = "Backend for sátni.org"
authors = ["Børre Gaup <albbas@gmail.com>"]
license = "GPLv3+"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
lxml = "4.9.3"
termwikiimporter = {path = "../../langtech/tools/TermWikiImporter"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.195.3"}
satni-back-python = {path = "../satni-back-rust/satni-back-python"}
uvicorn = {extras = ["standard"], version = "^0.23.1"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.280"
nose2 = "^0.13.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
  "I",   # isort
]


